Prepare the feature check for MC1.10

+++ src/com/dre/brewery/Brew.java
@@ -280,7 +280,7 @@ public class Brew {
 
 	// Do some regular updates
 	public void touch() {
-		lastUpdate = (int) ((float) (System.currentTimeMillis() - installTime) / 3600000F);
+		lastUpdate = (int) ((double) (System.currentTimeMillis() - installTime) / 3600000D);
 	}
 
 	public int getDistillRuns() {
